Extract webrtc logics to a custom hook
